// src/app/general-module/componentes-comunes/interfaces/Audit.inteface.ts
export interface AuditInteface {

    numeroExpediente: string;
    ubicacion: string;
    estadoExpediente: string;
    nombramiento: string;
    fechaPeriodoDel: Date;
    fechaPeriodoAl: Date;
    estadoNombramiento: string;
    nombrePlan: string;
    multa: number;
    ajuste: number;
    comentario: string;
}